Move get_second_element from copy.hpp to result_types.hpp
